chore: publish v1 of bentobox/trident/sushixswap/furo
