+++ .changeset/config.json
@@ -24,17 +24,13 @@
     "@sushiswap/token-price-api",
     "@sushiswap/cli",
     "@sushiswap/client",
-    "@sushiswap/database",
     "@sushiswap/react-query",
     "@sushiswap/redux-token-lists",
     "@sushiswap/redux-localstorage",
-    "@sushiswap/router",
-    "@sushiswap/token-lists",
     "@sushiswap/ui",
     "@sushiswap/wagmi",
     "@sushiswap/aggregator-job",
     "@sushiswap/pool-job",
-    "@sushiswap/furo",
     "@sushiswap/route-processor"
   ]
 }
