+++ apps/_root/package.json
@@ -26,7 +26,7 @@
     "@tanstack/react-table": "8.5.0",
     "@sushiswap/abi": "workspace:*",
     "@sushiswap/amm": "workspace:*",
-    "@sushiswap/bentobox": "workspace:*",
+    "@sushiswap/bentobox": "1.0.1",
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/currency": "workspace:*",
     "@sushiswap/database": "workspace:*",
@@ -43,9 +43,9 @@
     "@sushiswap/router-config": "workspace:*",
     "@sushiswap/stargate": "workspace:*",
     "@sushiswap/sushiswap": "workspace:*",
-    "@sushiswap/sushixswap": "workspace:*",
+    "@sushiswap/sushixswap": "1.0.1",
     "@sushiswap/tines": "workspace:*",
-    "@sushiswap/trident": "workspace:*",
+    "@sushiswap/trident": "1.0.1",
     "@sushiswap/types": "workspace:*",
     "@sushiswap/validate": "workspace:*",
     "@sushiswap/wagmi": "workspace:*",
