+++ apps/earn/package.json
@@ -28,7 +28,7 @@
     "@reduxjs/toolkit": "1.9.1",
     "@sushiswap/abi": "workspace:*",
     "@sushiswap/amm": "workspace:*",
-    "@sushiswap/bentobox": "workspace:*",
+    "@sushiswap/bentobox": "1.0.1",
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/client": "workspace:*",
     "@sushiswap/currency": "workspace:*",
@@ -43,7 +43,7 @@
     "@sushiswap/sushiswap": "workspace:*",
     "@sushiswap/token-lists": "workspace:*",
     "@sushiswap/route-processor": "workspace:*",
-    "@sushiswap/trident": "workspace:*",
+    "@sushiswap/trident": "1.0.1",
     "@sushiswap/types": "workspace:*",
     "@sushiswap/ui": "workspace:*",
     "@sushiswap/v3-sdk": "workspace:*",
