+++ apps/furo/package.json
@@ -31,12 +31,12 @@
     "@hookform/resolvers": "2.9.10",
     "@reduxjs/toolkit": "1.9.1",
     "@sushiswap/abi": "workspace:*",
-    "@sushiswap/bentobox": "workspace:*",
+    "@sushiswap/bentobox": "1.0.1",
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/currency": "workspace:*",
     "@sushiswap/dexie": "workspace:*",
     "@sushiswap/format": "workspace:*",
-    "@sushiswap/furo": "workspace:*",
+    "@sushiswap/furo": "1.0.1",
     "@sushiswap/graph-client": "workspace:*",
     "@sushiswap/hooks": "workspace:*",
     "@sushiswap/log": "workspace:*",
