+++ config/viem/CHANGELOG.md
@@ -1,0 +1,7 @@
+# @sushiswap/viem-config
+
+## 0.0.2
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
