+++ jobs/pool/package.json
@@ -57,7 +57,7 @@
     "@graphql-mesh/runtime": "0.46.22",
     "@graphql-mesh/cross-helpers": "0.3.4",
     "@graphql-mesh/store": "0.9.20",
-    "@sushiswap/bentobox": "workspace:*",
+    "@sushiswap/bentobox": "1.0.1",
     "@sushiswap/tines": "workspace:*",
     "@sushiswap/abi": "workspace:*",
     "@sushiswap/chain": "workspace:*",
