+++ packages/abi/CHANGELOG.md
@@ -1,0 +1,7 @@
+# @sushiswap/abi
+
+## 0.0.1
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
