+++ packages/abi/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@sushiswap/abi",
-  "version": "0.0.0",
-  "private": true,
+  "version": "0.0.1",
   "description": "Sushi ABI",
   "keywords": [
     "sushi",
