+++ packages/amm/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sushiswap/amm",
-  "version": "0.1.1",
+  "version": "0.1.3",
   "description": "Sushi AMM",
   "keywords": [
     "sushi",
@@ -40,7 +40,7 @@
     "@sushiswap/currency": "workspace:*",
     "@sushiswap/math": "workspace:*",
     "@sushiswap/tines": "workspace:*",
-    "@sushiswap/trident": "workspace:*",
+    "@sushiswap/trident": "1.0.1",
     "tiny-invariant": "1.3.1",
     "tiny-warning": "1.0.3",
     "zod": "3.20.2"
