+++ packages/database/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@sushiswap/database",
-  "version": "0.0.0",
-  "private": true,
+  "version": "0.0.1",
   "description": "Sushi Database",
   "keywords": [
     "sushi",
