+++ packages/react-query/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@sushiswap/react-query",
   "version": "0.0.0",
-  "private": true,
   "description": "Sushi React Query",
   "keywords": [
     "sushi",
@@ -47,17 +46,13 @@
     "@ethersproject/address": "5.7.0",
     "@ethersproject/bignumber": "5.7.0",
     "@ethersproject/units": "^5.7.0",
-    "@sushiswap/abi": "workspace:*",
     "@sushiswap/amm": "workspace:*",
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/client": "workspace:*",
     "@sushiswap/currency": "workspace:*",
-    "@sushiswap/database": "workspace:*",
     "@sushiswap/dexie": "workspace:*",
     "@sushiswap/math": "workspace:*",
     "@sushiswap/types": "workspace:*",
-    "@sushiswap/ui": "workspace:*",
-    "@sushiswap/route-processor": "workspace:*",
     "@tanstack/react-query": "4.28.0",
     "@tanstack/react-query-devtools": "4.28.0",
     "@tanstack/react-query-persist-client": "4.28.0",
