+++ packages/router/CHANGELOG.md
@@ -1,0 +1,16 @@
+# @sushiswap/router
+
+## 0.0.1
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
+- Updated dependencies [074ec9ebc]
+  - @sushiswap/sushiswap@0.1.1
+  - @sushiswap/bentobox@1.0.1
+  - @sushiswap/database@0.0.1
+  - @sushiswap/trident@1.0.1
+  - @sushiswap/v3-sdk@0.0.1
+  - @sushiswap/abi@0.0.1
+  - @sushiswap/amm@0.1.3
+  - @sushiswap/viem-config@0.0.2
