+++ packages/router/package.json
@@ -1,7 +1,6 @@
 {
   "name": "@sushiswap/router",
-  "version": "0.0.0",
-  "private": true,
+  "version": "0.0.1",
   "description": "Sushi Router",
   "keywords": [
     "sushi",
@@ -43,13 +42,13 @@
     "@ethersproject/solidity": "5.7.0",
     "@sushiswap/abi": "workspace:*",
     "@sushiswap/amm": "workspace:*",
-    "@sushiswap/bentobox": "workspace:*",
+    "@sushiswap/bentobox": "1.0.1",
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/currency": "workspace:*",
     "@sushiswap/math": "workspace:*",
     "@sushiswap/tines": "workspace:*",
     "@sushiswap/sushiswap": "workspace:*",
-    "@sushiswap/trident": "workspace:*",
+    "@sushiswap/trident": "1.0.1",
     "@sushiswap/v3-sdk": "workspace:*",
     "@sushiswap/viem-config": "workspace:*",
     "ethers": "5.7.2",
