+++ packages/token-lists/package.json
@@ -27,40 +27,33 @@
   "jest": {
     "preset": "@sushiswap/jest-config/node"
   },
-  "devDependencies": {
+  "dependencies": {
     "@ethersproject/address": "5.7.0",
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/currency": "workspace:*",
+    "@sushiswap/math": "workspace:*",
+    "@uniswap/token-lists": "^1.0.0-beta.31",
+    "tiny-invariant": "1.3.1"
+  },
+  "devDependencies": {
     "@sushiswap/eslint-config": "workspace:*",
     "@sushiswap/jest-config": "workspace:*",
-    "@sushiswap/math": "workspace:*",
     "@sushiswap/prettier-config": "workspace:*",
     "@sushiswap/typescript-config": "workspace:*",
     "@types/jest": "^29.2.0",
-    "@uniswap/token-lists": "^1.0.0-beta.29",
+    "@types/node": "^18.0.0",
+    "@types/react": "^18.0.0",
     "eslint": "8.29.0",
     "jest": "29.3.1",
-    "tiny-invariant": "1.3.1",
+    "react": "18.2.0",
     "typescript": "5.0.4"
   },
   "peerDependencies": {
-    "@ethersproject/address": "5.7.0",
-    "@sushiswap/currency": "*",
-    "@sushiswap/math": "*",
-    "@uniswap/token-lists": "*"
+    "react": "18.2.0"
   },
   "peerDependenciesMeta": {
-    "@ethersproject/address": {
-      "optional": false
-    },
-    "@sushiswap/currency": {
-      "optional": false
-    },
-    "@sushiswap/math": {
-      "optional": false
-    },
-    "@uniswap/token-lists": {
-      "optional": false
+    "react": {
+      "optional": true
     }
   }
 }
