+++ packages/v2-sdk/CHANGELOG.md
@@ -1,0 +1,7 @@
+# @sushiswap/v2-sdk
+
+## 0.0.2
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
