+++ packages/v2-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sushiswap/v2-sdk",
-  "version": "0.0.0",
+  "version": "0.0.2",
   "private": true,
   "description": "SushiSwap V2 SDK",
   "keywords": [
@@ -41,8 +41,6 @@
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/currency": "workspace:*",
     "@sushiswap/math": "workspace:*",
-    "@sushiswap/tines": "workspace:*",
-    "@sushiswap/trident": "workspace:*",
     "tiny-invariant": "1.3.1",
     "tiny-warning": "1.0.3",
     "zod": "3.20.2"
