+++ packages/v3-sdk/CHANGELOG.md
@@ -1,0 +1,9 @@
+# @sushiswap/v3-sdk
+
+## 0.0.1
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
+- Updated dependencies [074ec9ebc]
+  - @sushiswap/amm@0.1.3
