+++ packages/v3-sdk/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sushiswap/v3-sdk",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "description": "Sushi V3 SDK",
   "keywords": [
@@ -47,7 +47,6 @@
     "@sushiswap/currency": "workspace:*",
     "@sushiswap/math": "workspace:*",
     "@sushiswap/tines": "workspace:*",
-    "@sushiswap/trident": "workspace:*",
     "tiny-invariant": "1.3.1",
     "tiny-warning": "1.0.3",
     "zod": "3.20.2"
