+++ packages/wagmi/package.json
@@ -30,13 +30,13 @@
     "@heroicons/react": "2.0.13",
     "@sushiswap/abi": "workspace:*",
     "@sushiswap/amm": "workspace:*",
-    "@sushiswap/bentobox": "workspace:*",
+    "@sushiswap/bentobox": "1.0.1",
     "@sushiswap/chain": "workspace:*",
     "@sushiswap/client": "workspace:*",
     "@sushiswap/currency": "workspace:*",
     "@sushiswap/dexie": "workspace:*",
     "@sushiswap/format": "workspace:*",
-    "@sushiswap/furo": "workspace:*",
+    "@sushiswap/furo": "1.0.1",
     "@sushiswap/gas": "workspace:*",
     "@sushiswap/hooks": "workspace:*",
     "@sushiswap/math": "workspace:*",
@@ -45,9 +45,9 @@
     "@sushiswap/router": "workspace:*",
     "@sushiswap/router-config": "workspace:*",
     "@sushiswap/sushiswap": "workspace:*",
-    "@sushiswap/sushixswap": "workspace:*",
+    "@sushiswap/sushixswap": "1.0.1",
     "@sushiswap/tines": "workspace:*",
-    "@sushiswap/trident": "workspace:*",
+    "@sushiswap/trident": "1.0.1",
     "@sushiswap/types": "workspace:*",
     "@sushiswap/ui": "workspace:*",
     "@sushiswap/v3-sdk": "workspace:*",
