+++ protocols/bentobox/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@sushiswap/bentobox",
-  "version": "0.5.0",
+  "version": "1.0.1",
   "description": "BentoBox",
   "keywords": [
     "bentobox",
