+++ protocols/furo/CHANGELOG.md
@@ -1,0 +1,21 @@
+# @sushiswap/furo
+
+## 1.0.1
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
+- Updated dependencies [074ec9ebc]
+  - @sushiswap/sushiswap@0.1.1
+  - @sushiswap/bentobox@1.0.1
+
+## 1.0.0
+
+### Major Changes
+
+- release
+
+### Patch Changes
+
+- Updated dependencies
+  - @sushiswap/bentobox@1.0.0
