+++ protocols/sushiswap/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sushiswap/sushiswap
 
+## 0.1.1
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
+
 ## 0.1.0
 
 ### Minor Changes
