+++ protocols/sushixswap/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sushiswap/sushixswap
 
+## 1.0.1
+
+### Patch Changes
+
+- 074ec9ebc: initial publish chunk
+
+## 1.0.0
+
+### Major Changes
+
+- release
+
 ## 0.1.0
 
 ### Minor Changes
