+++ protocols/trident/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @sushiswap/bentobox
+# @sushiswap/trident
 
 ## 1.0.1
 
@@ -11,9 +11,3 @@
 ### Major Changes
 
 - release
-
-## 0.5.0
-
-### Minor Changes
-
-- 82c96649: initial release
